requirements.txt
# Core web framework
fastapi>=0.110.0

# ASGI server
uvicorn[standard]>=0.29.0
ngrok

# HTTP client (used in telegram_bot.py for webhook registration)
httpx>=0.27.0

# Environment variable loading
python-dotenv>=1.0.0

# ORM + PostgreSQL driver
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Discord gateway
discord.py>=2.3.2

# Discord Ed25519 signature verification
PyNaCl>=1.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
